dddd

+++ project/src/main/java/com/ourproject/portfoliotracker/data/portfolio/PortfolioService.java
@@ -1,10 +1,14 @@
 package com.ourproject.portfoliotracker.data.portfolio;
 
-import com.ourproject.portfoliotracker.data.portfolio.PortfolioEntity;
-import com.ourproject.portfoliotracker.data.portfolio.PortfolioRepository;
+import com.ourproject.portfoliotracker.data.account.AccountRepository;
+
+import org.modelmapper.ModelMapper;
+import org.modelmapper.TypeToken;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.lang.reflect.Type;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
@@ -13,47 +17,85 @@ import java.util.stream.StreamSupport;
 public class PortfolioService {
 
     private final PortfolioRepository portfolioRepository;
+    private final AccountRepository accountRepository;
+    private Integer pageSize;
 
-    public PortfolioService(PortfolioRepository portfolioRepository) {
+    @Autowired
+    public PortfolioService(PortfolioRepository portfolioRepository, AccountRepository accountRepository) {
         this.portfolioRepository = portfolioRepository;
+        this.accountRepository = accountRepository;
     }
 
-    //Creating an portfolio object
-    //Returning persisted object
-    public PortfolioEntity addPortfolio(PortfolioEntity portfolio) {
+    public PortfolioEntity addPortfolio(PortfolioDTO portfolioDTO) {
+        PortfolioEntity portfolio = new PortfolioEntity();
+        portfolio.setName(portfolioDTO.getName());
+        portfolio.setDescription(portfolioDTO.getDescription());
         return portfolioRepository.save(portfolio);
     }
 
+    //PortfolioEntity ->
+    public List<PortfolioDTO> getFirst20Portfolio(String userName, Integer pageId) {
+        Integer id = accountRepository.findByUserName(userName).getAccountId();
+        List<PortfolioEntity> portfolioEntities = portfolioRepository.findAllByAccountId(id);
+        if (portfolioEntities == null) {
+            throw new RuntimeException("Portfolios of id " + id + " not found");
+        }
+
+        ModelMapper modelMapper = new ModelMapper();
+        Type listType = new TypeToken<List<PortfolioDTO>>(){}.getType();
+        List<PortfolioDTO> portfolioDTOS = modelMapper.map(portfolioEntities, listType);
+
+        if (pageId < 1) {
+            throw new RuntimeException("pageId cannot be negative or zero: " + pageId);
+        }
+
+        return portfolioDTOS.subList(20 * (pageId - 1), 20 * pageId + 1);
+    }
+
     //Returning a List of all portfolios
-    public List<PortfolioEntity> getAllPorfolios() {
+ /*   public List<PortfolioEntity> getAllPorfolios() {
         return StreamSupport
                 .stream(portfolioRepository.findAll().spliterator(), false)
                 .collect(Collectors.toList());
-    }
+    }*/
 
-    //Getting a Portfolio by id
-    public PortfolioEntity getPorfolio(Integer accountId, String name) {
+    public PortfolioDTO getPortfolio(Integer accountId, String name) {
         PortfolioEntity portfolio = portfolioRepository.findByAccountIdAndName(accountId, name);
         if (portfolio == null) {
             throw new RuntimeException("Portfolio of id " + accountId + " and name " + name + " not found");
         }
+        PortfolioDTO portfolioDTO = new PortfolioDTO();
+        portfolioDTO.setName(portfolio.getName());
+        portfolioDTO.setDescription(portfolio.getDescription());
+        return portfolioDTO;
+    }
+
+    public PortfolioEntity deletePortfolio(Integer accountId, String name) {
+        PortfolioEntity portfolio = portfolioRepository.findByAccountIdAndName(accountId, name);
+        if (portfolio == null) {
+            throw new RuntimeException("Portfolio of id " + accountId + " and name " + name + " not found");
+        }
+        portfolioRepository.delete(portfolio);
         return portfolio;
     }
 
-    //Deleting a Portfolio by id
-/*    public PortfolioEntity deletePorfolio(Integer id) {
-        //Validating, if the object does not exist it will throw an error
-        PortfolioEntity portfolio = getPorfolio(id);
-        portfolioRepository.delete(portfolio);
+    @Transactional
+    public PortfolioEntity editName(Integer accountId, String portfolioName, String newName) {
+        PortfolioEntity portfolio = portfolioRepository.findByAccountIdAndName(accountId, portfolioName);
+        if (portfolio == null) {
+            throw new RuntimeException("Portfolio of id " + accountId + " and name " + portfolioName + " not found");
+        }
+        portfolio.setName(newName);
         return portfolio;
-    }*/
+    }
 
-/*    @Transactional
-    public PortfolioEntity editPorfolio(Integer id,PortfolioEntity portfolioObj) {
-        PortfolioEntity portfolio = getPorfolio(id);
-        portfolio.setAccountId(portfolioObj.getAccountId());
-        portfolio.setName(portfolioObj.getName());
-        portfolio.setDescription(portfolioObj.getDescription());
+    @Transactional
+    public PortfolioEntity editDescription(Integer accountId, String portfolioName, String newDescription) {
+        PortfolioEntity portfolio = portfolioRepository.findByAccountIdAndName(accountId, portfolioName);
+        if (portfolio == null) {
+            throw new RuntimeException("Portfolio of id " + accountId + " and name " + portfolioName + " not found");
+        }
+        portfolio.setDescription(newDescription);
         return portfolio;
-    }*/
+    }
 }
